requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
Werkzeug==3.0.1
flask-limiter==3.10.0  # Updated to a version that exists
email-validator==2.1.0
python-dateutil==2.8.2
flask-limiter==3.3.3